requirements.txt
pip
ipython
ipdb
wheel
python-poppler
numpy
pandas
matplotlib
google-ngram-downloader
tensorflow
